models/softs-new/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "softs-new"
version = "0.1.0"
description = "Standalone Kubeflow-native SOFTS C-MAPSS research workflows"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
  "mambasl-new",
  "einops>=0.8,<0.9",
  "kfp[kubernetes]==2.15.0",
  "kubernetes>=29,<31",
  "numpy>=1.24,<2",
  "optuna>=4.5,<5",
  "pandas>=2.0,<2.3",
  "pydantic>=2.7,<3",
  "PyYAML>=6,<7",
  "requests>=2.31,<3",
  "scikit-learn>=1.3,<1.6",
  "torch>=2.4,<2.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[project.scripts]
softs-new = "softs_new.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
